requirements.txt
requests
sparql-client==1.3